Initial support for opening files other than editor.js
Editing started to get cumbersome with such a long file (right now,
editor.js has 407 lines and it's hard to find stuff), so I needed to
either implement Find or Open. I chose Open as it will allow me to also
edit index.html and style.css so I don't have to open vim to edit them.

(My Ctrl key is starting to break, so this will be a welcome change.)

For this I needed to add a prompt element (since quick and dirty
prompt() is not supported in electron), and reorganize the editor DOM
structure a bit. Also using flexbox now instead of absolutely positioned
elements.

This is a temporary commit before trying it out to actually edit stuff,
since it's likely to mess up with my files.